refactor(web): extract the model setting controls into a shared module
The slider, the setting row, and the reasoning stop constants were private to
the chat model selector. The admin provider modal needs the same controls, so
move them out rather than grow a second implementation.

Pure move. maxSupportedReasoningStop stays behind as a one-line adapter over
maxReasoningStop, since the admin side holds a raw ModelConfiguration rather
than an LLMOption.
