+++ web/src/sections/model-selector/setting-controls.tsx
@@ -1,0 +1,169 @@
+"use client";
+
+import * as SliderPrimitive from "@radix-ui/react-slider";
+import { Text, Tooltip } from "@opal/components";
+import { Section } from "@opal/layouts";
+import { cn } from "@opal/utils";
+import type { IconFunctionComponent } from "@opal/types";
+import { Disabled } from "@opal/core";
+import { ReasoningEffortOverride } from "@/lib/languageModels/types";
+
+/** The levels every reasoning model supports, in ascending order. */
+export const BASE_REASONING_STOPS: ReasoningEffortOverride[] = [
+  "off",
+  "low",
+  "medium",
+  "high",
+];
+
+/** Every stop the slider renders. Unsupported stops are greyed, never hidden. */
+export const ALL_REASONING_STOPS: ReasoningEffortOverride[] = [
+  ...BASE_REASONING_STOPS,
+  "xhigh",
+];
+
+export const REASONING_STOP_LABELS: Record<ReasoningEffortOverride, string> = {
+  off: "Off",
+  low: "Low",
+  medium: "Medium",
+  high: "High",
+  xhigh: "XHigh",
+};
+
+export const UNSUPPORTED_SETTING_TOOLTIP =
+  "Modifying this setting is not supported for this model.";
+
+export const UNKNOWN_CONTEXT_TOOLTIP =
+  "Context size is not available for this model. Chats still apply a token limit automatically.";
+
+/** Highest supported stop. Nothing from an older backend means the base set. */
+export function maxReasoningStop(
+  supported: ReasoningEffortOverride[] | undefined
+): number {
+  if (!supported) return BASE_REASONING_STOPS.length - 1;
+  return Math.max(
+    -1,
+    ...supported.map((effort) => ALL_REASONING_STOPS.indexOf(effort))
+  );
+}
+
+export function reasoningStopIndex(
+  effort: ReasoningEffortOverride | null | undefined
+): number {
+  return effort ? ALL_REASONING_STOPS.indexOf(effort) : -1;
+}
+
+export function formatContextWindow(tokens: number): string {
+  if (tokens >= 1_000_000)
+    return `${(tokens / 1_000_000).toFixed(1).replace(/\.0$/, "")}M`;
+  return tokens >= 1000 ? `${Math.round(tokens / 1000)}K` : `${tokens}`;
+}
+
+const SLIDER_THUMB_CLASS =
+  "block size-3 rounded-full bg-background-neutral-00 shadow-[0_0_2px_1px_rgba(0,0,0,0.15)] focus:outline-none";
+const SLIDER_TRACK_CLASS =
+  "h-1.5 w-full overflow-hidden rounded bg-background-tint-02";
+const SLIDER_FILL_CLASS = "h-full bg-theme-primary-05";
+
+interface PaneSliderProps {
+  value: number;
+  min: number;
+  max: number;
+  step: number;
+  disabled?: boolean;
+  onValueChange: (value: number) => void;
+  onValueCommit: (value: number) => void;
+}
+
+export function PaneSlider({
+  value,
+  min,
+  max,
+  step,
+  disabled,
+  onValueChange,
+  onValueCommit,
+}: PaneSliderProps) {
+  return (
+    <SliderPrimitive.Root
+      className="relative flex h-7 w-full cursor-pointer touch-none select-none items-center"
+      value={[value]}
+      min={min}
+      max={max}
+      step={step}
+      disabled={disabled}
+      onValueChange={(vals) => vals[0] !== undefined && onValueChange(vals[0])}
+      onValueCommit={(vals) => vals[0] !== undefined && onValueCommit(vals[0])}
+    >
+      <SliderPrimitive.Track
+        className={cn(SLIDER_TRACK_CLASS, "relative grow")}
+      >
+        <SliderPrimitive.Range className={cn(SLIDER_FILL_CLASS, "absolute")} />
+      </SliderPrimitive.Track>
+      <SliderPrimitive.Thumb className={SLIDER_THUMB_CLASS} />
+    </SliderPrimitive.Root>
+  );
+}
+
+interface SettingRowProps {
+  icon: IconFunctionComponent;
+  title: string;
+  value?: string;
+  /** Shown when hovering the value readout. */
+  valueTooltip?: string;
+  caption: string;
+  disabled?: boolean;
+  /** Shown when hovering the row while disabled. */
+  disabledTooltip?: string;
+  children?: React.ReactNode;
+}
+
+export function SettingRow({
+  icon: Icon,
+  title,
+  value,
+  valueTooltip,
+  caption,
+  disabled = false,
+  disabledTooltip,
+  children,
+}: SettingRowProps) {
+  return (
+    <Disabled disabled={disabled} tooltip={disabledTooltip} tooltipSide="top">
+      <Section
+        alignItems="stretch"
+        height="auto"
+        gap={0}
+        padding={1.5}
+        className="rounded-08"
+      >
+        <Section
+          flexDirection="row"
+          justifyContent="between"
+          height="auto"
+          gap={2}
+        >
+          <Section flexDirection="row" width="fit" height="auto" gap={2}>
+            <Section width={1.25} height={1.25} className="text-text-04">
+              <Icon size={16} />
+            </Section>
+            <Text font="main-ui-action">{title}</Text>
+          </Section>
+          {value !== undefined && (
+            <Tooltip tooltip={valueTooltip} side="top">
+              <Text font="secondary-mono" color="text-04">
+                {value}
+              </Text>
+            </Tooltip>
+          )}
+        </Section>
+        {children}
+        <Section alignItems="stretch" height="auto" className="mt-2">
+          <Text font="secondary-body" color="text-03">
+            {caption}
+          </Text>
+        </Section>
+      </Section>
+    </Disabled>
+  );
+}
